fossid-webapp: Add a property to keep failed scans
When there is an error during a FossID scan, the default behavior is to
delete the scan to avoid cluttering the FossID instance. While useful,
this makes error analysis of runs difficult. Therefore, this commit adds
an option to keep failed scans.

+++ scanner/src/main/kotlin/scanners/fossid/FossId.kt
@@ -342,9 +342,11 @@ class FossId internal constructor(
                     val scanResult = ScanResult(provenance, details, summary)
                     results.getOrPut(pkg) { mutableListOf() } += scanResult
 
-                    createdScans.forEach { code ->
-                        log.warn { "Deleting scan '$code' during exception cleanup." }
-                        deleteScan(code)
+                    if (!config.keepFailedScans) {
+                        createdScans.forEach { code ->
+                            log.warn { "Deleting scan '$code' during exception cleanup." }
+                            deleteScan(code)
+                        }
                     }
                 }
             }
